project/pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "cloudpickle>=3.1.2",
    "jsonschema>=4.25.1",
    "numpy>=2.3.5",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.1",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "toml>=0.10.2",
]

[tool.setuptools.packages.find]
include = ["sql_rag*", "doc_rag*", "common*"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
